Change to GenderDetector
After @davidcelis rightfully pointed out it wouldn't be unusual if
someone had their own Gender class.
