+++ .ruby-gemset
@@ -1,1 +1,1 @@
-gender-detector
+gender_detector
