+++ README.rdoc
@@ -1,10 +1,10 @@
-= Gender::Detector 
+= GenderDetector
 {<img src="https://secure.travis-ci.org/bmuller/sexmachine.png?branch=master" alt="Build Status" />}[https://travis-ci.org/bmuller/sexmachine]
 
 This gem uses the underlying data from the program "gender" by Jorg Michael (described {here}[http://www.autohotkey.com/community/viewtopic.php?t=22000]).  Its use is pretty straightforward:
 
-  >> require 'gender/detector'
-  >> d = Gender::Detector.new
+  >> require 'gender_detector'
+  >> d = GenderDetector.new
   >> d.get_gender("Bob")
   :male
   >> d.get_gender("Sally")
@@ -26,15 +26,15 @@ Additionally, you can give preference to specific countries:
   >> d.get_gender("Jamie", :great_britain)
    => :mostly_male
 
-If you have an alterative data file, you can pass that in as an optional filename argument to the Detector.  Additionally, you can create a detector that is not case sensitive (default *is* to be case sensitive):
+If you have an alterative data file, you can pass that in as an optional filename argument to the GenderDetector.  Additionally, you can create a detector that is not case sensitive (default *is* to be case sensitive):
 
-  >> d = Gender::Detector.new(:case_sensitive => false)
+  >> d = GenderDetector.new(:case_sensitive => false)
   >> d.get_gender "sally"
    => :female
   >> d.get_gender "Sally"
    => :female
 
-Try to avoid creating many Detectors, as each creation means reading in the data file.
+Try to avoid creating many GenderDetectors, as each creation means reading in the data file.
 
 = Licenses
 The genderator code is distributed under the GPLv3.  The data file nam_dict.txt is released under the GNU Free Documentation License.
