+++ Rakefile
@@ -5,7 +5,7 @@ require 'rake/testtask'
 task :default => [:test]
 
 RDoc::Task.new("doc") { |rdoc|
-  rdoc.title = "Gender::Detector - Detect gender from first name"
+  rdoc.title = "GenderDetector - Detect gender from first name"
   rdoc.rdoc_dir = 'docs'
   rdoc.rdoc_files.include('README.rdoc')
   rdoc.rdoc_files.include('lib/**/*.rb')
