+++ lib/gender_detector/version.rb
@@ -1,0 +1,13 @@
+class GenderDetector
+  class Version
+    MAJOR = 0
+    MINOR = 1
+    PATCH = 0
+
+    def self.to_s
+      [MAJOR, MINOR, PATCH].compact.join('.')
+    end
+  end
+
+  VERSION = Version.to_s
+end
