+++ test/gender_detector_test.rb
@@ -2,11 +2,11 @@
 require 'minitest/autorun'
 lib = File.expand_path('../../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'gender/detector'
+require 'gender_detector'
 
 class GenderDetectorTest < MiniTest::Test
   def setup
-    @d = Gender::Detector.new
+    @d = GenderDetector.new
   end
 
   def test_get_gender
@@ -26,13 +26,13 @@ class GenderDetectorTest < MiniTest::Test
   end
 
   def test_case_insensitivity
-    d = Gender::Detector.new(:case_sensitive => false)
+    d = GenderDetector.new(:case_sensitive => false)
     assert_equal :female, d.get_gender("sally")
     assert_equal :female, d.get_gender("Sally")
   end
 
   def test_name_exists
-    d = Gender::Detector.new(:case_sensitive => false)
+    d = GenderDetector.new(:case_sensitive => false)
     assert d.name_exists?("Sally")
     assert d.name_exists?("Carlos")
     assert d.name_exists?("Rosario")
